+++ uav_interface/SRFSonar.h
@@ -37,7 +37,7 @@ class SRFSonar : public Device
 
   private:
     uint16_t *_lastReadingRaw;
-    uint16_t _lastReadingRawLength;
+    uint32_t _lastReadingRawLength;
     uint8_t _I2CAddress;
     uint8_t _startAddress;
     uint8_t _nrOfBytesToRead;
